refactor: refactor types to simplify readability
